Extract a few components
